Unit Tests for GetReceipts() and GetReceipts(DateTime)
